requirements.txt
albumentations==1.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.1
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
executing==2.2.1
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
hub-sdk==0.0.24
huggingface_hub==1.7.1
idna==3.11
ImageIO==2.37.3
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
lightning==2.6.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
lazy-loader==0.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==1.26.4
opencv-python==4.11.0.86
opencv-python-headless==4.9.0.80
packaging==26.0
pandas==3.0.1
parso==0.8.6
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.9.4
prompt_toolkit==3.0.52
protobuf==6.33.6
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycocotools==2.0.7
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
pyzmq==27.1.0
qudida==0.0.4
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
sentry-sdk==2.55.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
stack-data==0.6.3
supervision==0.18.0
sympy==1.13.1
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2026.3.3
timm==1.0.25
torch==2.6.0
torchvision==0.21.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.1.0
urllib3==2.6.3
wandb==0.25.1
wcwidth==0.6.0
